requirements.txt
uvicorn
fastapi
jinja2
httpx
python-dotenv
openai
annotated-types==0.7.0
beautifulsoup4==4.13.3
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.8.0
defusedxml==0.7.1
et_xmlfile==2.0.0
fonttools==4.56.0
fpdf==1.7.2
fpdf2==2.8.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
Markdown==3.7
markdown2==2.5.3
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydyf==0.11.0
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
rsa==4.9
six==1.17.0
soupsieve==2.6
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
zopfli==0.2.3.post1
python-multipart